Remove extension from repo.
