--- src/common/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "references": [],
-  "compilerOptions": {
-    "module": "commonjs",
-    "lib": [],
-    "rootDir": ".",
-    "outDir": "../../lib/common",
-  },
-  "include": ["**/*"],
-}
