+++ src/disposable.ts
@@ -1,5 +1,5 @@
 import { Unsubscribable } from 'rxjs'
-import { Logger } from '../common/logging'
+import { Logger } from './logging'
 
 export interface Disposable {
     dispose(): void
