+++ src/logging.ts
@@ -1,3 +1,5 @@
+import { MessageConnection } from 'vscode-jsonrpc'
+import { LogMessageNotification } from 'vscode-languageserver-protocol'
 import { inspect } from 'util'
 import { MessageType } from 'vscode-languageserver-protocol'
 
@@ -91,6 +93,29 @@ export class MultiLogger extends AbstractLogger {
     protected logType(type: LogLevel, values: unknown[]): void {
         for (const logger of this.loggers) {
             logger[type](...values)
+        }
+    }
+}
+
+/**
+ * A logger implementation that sends window/logMessage notifications to an LSP client
+ */
+export class LSPLogger extends AbstractLogger {
+    /**
+     * @param client The client to send window/logMessage notifications to
+     */
+    constructor(private client: MessageConnection) {
+        super()
+    }
+
+    protected logType(type: LogLevel, values: unknown[]): void {
+        try {
+            this.client.sendNotification(LogMessageNotification.type, {
+                type: LOG_LEVEL_TO_LSP[type],
+                message: values.map(format).join(' '),
+            })
+        } catch (err) {
+            // ignore
         }
     }
 }
