--- src/server/cancellation.ts
@@ -1,8 +1,0 @@
-import axios, { CancelToken } from 'axios'
-import { CancellationToken } from 'vscode-jsonrpc'
-
-export function toAxiosCancelToken(token: CancellationToken): CancelToken {
-    const source = axios.CancelToken.source()
-    token.onCancellationRequested(() => source.cancel())
-    return source.token
-}
