--- src/server/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "references": [{ "path": "../common/tsconfig.json" }],
-  "compilerOptions": {
-    "module": "commonjs",
-    "lib": ["es2018", "esnext.asynciterable", "dom"],
-    "typeRoots": ["../node_modules/@types", "./src/types"],
-    "outDir": "../../lib/server",
-    "rootDir": ".",
-  },
-  "include": ["**/*"],
-}
