+++ src/tsconfig.ts
@@ -3,10 +3,10 @@ import { writeFile } from 'mz/fs'
 import { Span, Tracer } from 'opentracing'
 import { fileURLToPath, URL } from 'url'
 import { CancellationToken } from 'vscode-jsonrpc'
-import { throwIfCancelled } from '../common/cancellation'
-import { flatMapConcurrent } from '../common/ix'
-import { Logger } from '../common/logging'
-import { logErrorEvent, tracePromise } from '../common/tracing'
+import { throwIfCancelled } from './cancellation'
+import { flatMapConcurrent } from './ix'
+import { Logger } from './logging'
+import { logErrorEvent, tracePromise } from './tracing'
 import { ResourceRetrieverPicker } from './resources'
 
 export async function sanitizeTsConfigs({
