+++ src/yarn.ts
@@ -6,9 +6,9 @@ import { combineLatest, concat, fromEvent, merge, Unsubscribable } from 'rxjs'
 import { filter, map, switchMap, withLatestFrom } from 'rxjs/operators'
 import { Readable } from 'stream'
 import { CancellationToken, Disposable } from 'vscode-jsonrpc'
-import { createAbortError, throwIfCancelled } from '../common/cancellation'
-import { Logger, NoopLogger } from '../common/logging'
-import { tracePromise } from '../common/tracing'
+import { createAbortError, throwIfCancelled } from './cancellation'
+import { Logger, NoopLogger } from './logging'
+import { tracePromise } from './tracing'
 import { disposeAll } from './disposable'
 import { Progress, ProgressProvider } from './progress'
